pyproject.toml
[tool.poetry]
name = "comicbookapi"
version = "0.1.0"
description = ""
authors = ["Artem <kznartee16reg@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
django = "4.1.7"
djangorestframework = "^3.14.0"
requests = "^2.28.2"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
